Use zwp_linux_dmabuf_feedback_v1 tranche_formats
zwp_linux_dmabuf_v1 version 4 introduced the concept of dmabuf feedback.
This advertises supported formats and modifiers in so-called tranches,
allowing multiple devices and configurations to be advertised.

The global format and modifier events were deprecated in the same
protocol revision, so a client will receive either format and modifier
events on the zwp_linux_dmabuf_v1 object, or have access to a
zwp_linux_dmabuf_feedback_v1 object, never both.

7d050204c6ca ("Implement dmabuf feedback main_device in v4") switched to
v4 to get access to the main_device advertisement from the new feedback
objects, but did not implement support for tranche formats and did not
account for the old events going silent.

Fix the regression by implementing support for tranche_formats. We are
not yet implementing tranche_flags, so regular and scanout format and
modifier pairs remain mixed.

Bug: 437811360
Change-Id: I9fa0018aa7ec11559770e00ca792c15280e3edd5